data/requirements_api.txt
flask==2.3.3
avro-python3==1.10.2
requests==2.31.0
